Drop NumPy build dep from cudf (keep in pylibcudf)
With the pylibcudf refactor, only pylibcudf looks for NumPy headers.
Whereas the Python cudf package does not. Given this, we shouldn't need
the NumPy build dependency on anything other than pylibcudf. So drop the
NumPy build dependency elsewhere. In addition to cleaning up the
dependencies this will ease the NumPy 2 upgrade effort.