Use NumPy 1 in legacy Pandas testing
These Pandas test now try all minor versions of Pandas 2 before Pandas
latest. Though only Pandas 2.2.2+ is NumPy 2. However when we downgrade
Pandas, NumPy is not downgraded to NumPy 1. As a result, we get the
expected errors when trying to mix incompatible NumPy & Pandas versions.
This is in part due to Pandas not having an upper bound on its NumPy
constraint. Though it is possible `pip` would ignore this anyways. So to
fix this issue, just use NumPy 1 for these older versions of Pandas,
which should work consistently for them. We have plenty of NumPy 2
coverage outside of this particular test. Especially consider we do run
tests with the latest Pandas and NumPy first, which will use NumPy 2 and
Pandas 2.2.2+.


https://pandas.pydata.org/pandas-docs/version/2.2.2/whatsnew/v2.2.2.html#pandas-2-2-2-is-now-compatible-with-numpy-2-0

https://github.com/pandas-dev/pandas/blob/v2.0.3/pyproject.toml#L26-L28

https://github.com/pandas-dev/pandas/blob/v2.1.4/pyproject.toml#L32-L34