Need to check for nulls in nested dtypes
To do this we need to inspect the polars datatypes, since by the time
we've converted to the pylibcudf one, the nested element types have
been lost.

We don't do this eagerly during dtype conversion because we still want
to allow scalar literals with null dtype that will then be cast to a
non-null dtype.