Fix order-preservation in cudf-polars groupby
When we are requested to maintain order in groupby aggregations we
must post-process the result by computing a permutation between the
wanted order (of the input keys) and the order returned by the groupby
aggregation. To do this, we can perform a join between the two unique
key tables. Previously, we assumed that the gather map returned in
this join for the left (wanted order) table was the identity. However,
this is not guaranteed, in addition to computing the match between the
wanted key order and the key order we have, we must also apply the
permutation between the left gather map order and the identity.

- Closes #16893