Fix `host_span` constructor to correctly copy `is_device_accessible` (#17020)
One of the `host_span` constructors was not updated when we added `is_device_accessible`, so the value was not assigned. 
This PR fixes this simple error and adds tests that checks that this property is correctly set when creating `host_span`s.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Muhammad Haseeb (https://github.com/mhaseeb123)

URL: https://github.com/rapidsai/cudf/pull/17020