Unify treatment of `Expr` and `IR` nodes in cudf-polars DSL (#17016)
As part of in-progress multi-GPU work, we will likely want to:

1. Introduce additional nodes into the `IR` namespace;
2. Implement rewrite rules for `IR` trees to express needed communication patterns;
3. Write visitors that translate expressions into an appropriate description for whichever multi-GPU approach we end up taking.

It was already straightforward to write generic visitors for `Expr` nodes, since those uniformly have a `.children` property for their dependents. In contrast, the `IR` nodes were more ad-hoc. To solve this, pull out the generic implementation from `Expr` into an abstract `Node` class. Now `Expr` nodes just inherit from this, and `IR` nodes do so similarly.

Redoing the `IR` nodes is a little painful because we want to make them hashable, so we have to provide a bunch of custom `get_hashable` implementations (the schema dict, for example, is not hashable).

With these generic facilities in place, we can now implement traversal and visitor infrastructure. Specifically, we provide:

- a mechanism for pre-order traversal of an expression DAG, yielding each unique node exactly once. This is useful if one wants to know if an expression contains some particular node;
- a mechanism for writing recursive visitors and then wrapping a caching scheme around the outside. This is useful for rewrites.

Some example usages are shown in tests.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)
  - Richard (Rick) Zamora (https://github.com/rjzamora)
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Murray (https://github.com/Matt711)
  - Richard (Rick) Zamora (https://github.com/rjzamora)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/17016