Turn off cudf.pandas 3rd party integrations tests for 24.12 (#17500)
Removes the third-party integration tests for the 24.12 nightly CI. We
need to do this to unblock CI. These tests have not been running
properly, and we just noticed that. There are more than a few failures
so we will have to resolve this in the next release. Future work is
tracked in #17490.