Force Thrust to use 32-bit offset type. (#17523)
This fixes the patch we use for Thrust to always get a 32-bit offset type. The net effect of this patch is that we are behaving as if `THRUST_FORCE_32_BIT_OFFSET_TYPE` is set. This replaces a previous patch which I mistakenly did not update between CCCL 2.6.x testing and 2.7.0-rc2 testing.

In the future we hope to configure this with CMake and drop the patches, but that will require us to use features from https://github.com/NVIDIA/cccl/pull/2844 (which is not available in 2.7.0-rc2).

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/17523