Added support for custom types in PTX parser (#17861)
The PTX parser currently enforces a signature for each parameter. 
It requires that the function have the following properties:

- The first parameter is an output pointer type, `void *`
- The other parameters are `void const *` pointers (if specified) or `int32` (if unspecified and unused)

As you can imagine, this is very limiting for any use beyond what it was initially designed for.

This merge request removes the constraint totally and lets the user specify the types of the parameters.
This is needed for more flexible code-gen in UDFs, i.e. defining a function that transforms multiple columns into a single output column using an index as the first parameter is impossible with this limitation.

Note that the register types and sizes are checked against the parameter types during the compilation process.
There's also only a minimal level of safety that can be guaranteed given we are translating from PTX back to CUDA and much of the type information is lost in the process.

Authors:
  - Basit Ayantunde (https://github.com/lamarrr)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Shruti Shivakumar (https://github.com/shrshi)

URL: https://github.com/rapidsai/cudf/pull/17861