Fix torch integration test (#17923)
Apart of #17490. Creating a torch tensor from a cudf.pandas proxy `Series` or `DataFrame` creates a device tensor if the underlying object is cudf. So this PR updates the `assert_eq` function in the module to convert the device tensor to a host tensor before comparing equality.

Authors:
  - Matthew Murray (https://github.com/Matt711)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/17923