Make `cudf.pandas` proxy array picklable (#17929)
Apart of #17490.

We employ custom pickling logic for our cudf.pandas wrapped types. The logic lets us serialize and de-serialize wrapped types by serializing and de-serializing the underlying wrapped types (ie. the type of `_fsproxy_wrapped`). This pickling logic is defined in `_FinalProxy`, which is the base class of all of our "final" proxy types. 

The failures in the integration tests occurred because this pickling logic wasn't used for the proxy numpy array type. This is because the "final" proxy array type inherits from an additional base class: `ProxyNDarrayBase` (which contains logic to inherit from `np.ndarray` ). And it comes before `_FinalProxy` in the classes MRO, so the custom pickling is not used.

Additionally, the custom pickling logic used for other proxy types is incompatible with our proxy array. So this PR defines a custom function for handling proxy array serialization.

Authors:
  - Matthew Murray (https://github.com/Matt711)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/17929