dask-cudf doc fixes
This updates the dask-cudf docs to fix a few things

1. Removes the majority of the API reference, which
   is mostly inherited from dask.dataframe.
2. Updates all references that previously pointed to
   our API docs to dasks (e.g. ``:meth:\`dataframe.read_parquet\``
   is now `:meth:\`dask.dataframe.read_parquet`` )
3. Links to best-practices (which was previously orphaned) from index.rst
4. fixes all remaining warnings

More context on the first change: our documentation
recommends using the ``dask.dataframe`` API with the cudf backend
set. This is functionally equivalent to using APIs from ``dask_cudf``
directly. I think the API refererence should match what we document,
and because we are inheriting all the docstrings there isn't really any
value in having them in both places (dask and dask-cudf).