Extend and simplify API for calculation of range-based rolling window offsets (#17807)
In both cudf.pandas and cudf-polars we would like to be able to use the existing libcudf functionality for computing the `preceding` and `following` columns for range-based (grouped) rolling windows.

The current functionality is designed with spark in mind and supports calculations with slightly different requirements compared to pandas and polars.

In this PR, we unify the construction of these offset column calculations to satisfy all uses.

Specifically:

1. We introduce a new window type: a `BOUNDED_OPEN` window. This is a range-based window where the endpoint of the range is _not_ included.
2. We add support for negative values of the window range offset parameter. The existing code only supports non-negative values.

The proposed interface for describing the requested window-type going forward is a strongly-typed one using `std::variant`. This removes the need for default-constructed scalars when specifying `UNBOUNDED` and `CURRENT_ROW` windows.

- Closes #17578 

**Performance improvements**

Spark permits nulls in the `orderby` column. In the grouped-rolling case, these nulls must be sorted at one end of each group. The current implementation finds the partition point in each group using `thrust::for_each` over each group and `thrust::partition_point` to find the break. For low-cardinality grouped rolling aggregations this can be very slow, since we have only a single thread searching over each group. We replace this with a segmented sum with CUB to find the per-group null count (and hence the break).

- Closes #17714

The dispatched functor for computing the bound given a row value has constexpr dispatch for the common, and required for pandas and polars, case that the `orderby` column has no-nulls. This shaves some milliseconds.

In the case that the `orderby` column does have nulls, we extend the interface such that the caller (who must have arranged that it is sorted correctly) tells us whether nulls were sorted `BEFORE` or `AFTER`. This avoids multiple kernel launches to deduce the null order, saving some milliseconds, and (in the grouped case) memory footprint. We polyfill this deduction so that the existing interface still works until we can deprecate and remove it.

**Guide for review**

The core logic is implemented in `range_utils.cuh`, specifically the `range_window_clamper`, dispatched to by `make_range_window`. To keep compile times under control, the template instantiation for computing preceding and following windows is moved into separate translation units: locally compilation takes ~10mins/TU. Things could be split out further if deemed necessary.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/17807