dask-polars: use splat everywhere. (#18243)
In the dask-graph, use the splat operator instead of list arguments everywhere. The splat operator is used in most places already, only a few functions are missing.
This makes it easier to transform the dask-graph after its creation, which will be useful for prototyping.

Authors:
  - Mads R. B. Kristensen (https://github.com/madsbk)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/cudf/pull/18243