Introduce ZSTD host-side compression and decompression APIs (#17935)
This PR uses [`libzstd`](https://github.com/facebook/zstd) to introduce ZSTD compression and decompression APIs to `libcudf`. 
We are able to test with `libzstd` in the conda environment since we get the library for 'free' as a dependency of `urllib3`; however the pip env requires `libzstd` to be already installed on the system. 
A separate PR will be opened for build changes to link to libzstd directly - please hold off on merging this PR until then.

Authors:
  - Shruti Shivakumar (https://github.com/shrshi)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Paul Mattione (https://github.com/pmattione-nvidia)
  - Vukasin Milovanovic (https://github.com/vuule)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/17935