Pass stream through when taking ownership from libcudf
When we take ownership of a table/column from libcudf we construct an
rmm DeviceBuffer via a DeviceBuffer's from_unique_ptr. This latter
takes a stream argument, defaulting to the default stream (the per
thread default stream if RMM was compiled with PTDS enabled). If RMM
determines that the stream we're taking ownership on is the default
stream, it syncs the stream.

In the newly stream-aware IO routines, by not providing the user's
stream when taking ownership of data from libcudf, we incur more
stream syncs than necessary. To avoid this, plumb through the explicit
user stream when taking ownership.