Use delete not free to release data allocated with new
When releasing nanoarrow schemas and arrays in pylibcudf interop we
allocate the space for the struct with new, but previously deallocated
it with free. This is technically incorrect: data allocated with new
should be deallocated with delete. To fix this, introduce a few more
inlined C++ functions that take the place of the calls to free.