Use delete not free to release data allocated with new (#18412)
When releasing nanoarrow schemas and arrays in pylibcudf interop we allocate the space for the struct with new, but previously deallocated it with free. This is technically incorrect: data allocated with new should be deallocated with delete. To fix this, introduce a few more inlined C++ functions that take the place of the calls to free.

Authors:
  - Lawrence Mitchell (https://github.com/wence-)
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/18412