Fix some duplicate toctree issues and improve groupby docs (#18580)
This PR addresses the problems in #18148 that can be fixed, namely those that were incorrect duplicates. In the process, I found that we were documenting groupby objects very differently from pandas, so I fixed that up to the best of my ability. The remaining "multiple toctree" warnings are not easily fixable because they are the result of duplicates that we have in place to match pandas documentation and the underlying autodoc commands are buried beneath autosummary layers. If we really want to fix then we need to add custom templates, which I don't think is worthwhile.

Closes #18148

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)

URL: https://github.com/rapidsai/cudf/pull/18580