Export cudf compiler flags and definitions (#18604)
This exports the compiler flags and definitions that are configured by libcudf, so the downstream applications can retrieve and use them directly. 

This is useful for the applications that heavily depend on libcudf and try to maintain a similar CMake build system as libcudf, such as `spark-rapids-jni`. They typically have to generate the same set of flags/definitions similar to what libcudf uses and have to periodically update their flags/definitions to sync with libcudf.

Authors:
  - Nghia Truong (https://github.com/ttnghia)

Approvers:
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)
  - Alessandro Bellina (https://github.com/abellina)

URL: https://github.com/rapidsai/cudf/pull/18604