Reduce register pressure for compute_column_kernel (#18226)
Fixes #17801 

This PR updates the AST operator to follow different code paths based on the presence of nulls and complex types in the input. This reduces register pressure and improves runtime performance when processing inputs that are not strings or decimals.

Authors:
  - https://github.com/matal-nvidia
  - Yunsong Wang (https://github.com/PointKernel)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)
  - Basit Ayantunde (https://github.com/lamarrr)
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - David Wendt (https://github.com/davidwendt)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/18226