Add support for large list host Arrow data conversion (#18562)
The Arrow specification supports both 32- and 64-bit offsets for list types. libcudf only supports the former, so up until now we have only supported that conversion. However, Polars regularly produces 64-bit list offsets. We need a way to ingest that data. In general, this cannot be done losslessly, but in the case where all of the data actually fits into the smaller list we can translate directly. This PR adds that functionality. Since it requires a copy, and since the use case is specifically targeted at cudf-polars, I have not added this support to the device interop layer, only the host interop.

Contributes to #18534

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Karthikeyan (https://github.com/karthikeyann)
  - Nghia Truong (https://github.com/ttnghia)
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/cudf/pull/18562