Fix host decompression of empty Snappy data (#18800)
issue #18445

Snappy compresses zero bytes of data with a single byte - a varint that encodes the uncompressed size as zero. No other data is in the compressed block in this case.
Host Snappy decompression does not take this case into account and fails with this one-byte input + empty output case.

This PR modifies the error checking in the Snappy decompression to account for this case. Also expanded the checks to better cover incorrectly sized input.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - David Wendt (https://github.com/davidwendt)

URL: https://github.com/rapidsai/cudf/pull/18800