Avoid RecursionError in custreamz test (#18887)
https://github.com/rapidsai/cudf/actions/runs/15119048978/job/42498435703?pr=18870#step:9:1722 observed a failure in a custreamz test.

I didn't get to a root cause as to why, but running the tests in a loop overnight did reproduce the failure. My best guess right now is "some weird interaction between streamz's `__del__` and distributed's `gc` callbacks". The easy solution for now is to disable distributed's gc callbacks for the duration of the tests.

Opening in draft to let this run locally for a while.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/18887