Fix tdigest percentile correctness.
This commit fixes tdigest percentile calculation correctness for cases low
row-counts.

The bug manifests when percentiles are calculated on a tdigest constructed
from say, 10 rows, for example.  The tdigest would then contain 10 centroids
all of equal weight (1.0).  When the percentile is calculated, one would
expect that the value of the rounded percentile would coincide exactly with
one of the rows.
E.g. Consider the following dataset:
```c++
auto const rows = { 708, 717, 769, 1022, 1097, 1108, 1400, 1460, 2469, 2761 };
```
One would expect `approx_percentile(rows, 0.2)` to equal the second row (`717`).
Instead, one sees the value `743`, i.e. the arithmetic mean of the second and
third rows.

This is the result of a quirk in how the diffs from the centroid mean is
calculated.  This has been rectified in this commit.

Signed-off-by: MithunR <mithunr@nvidia.com>