Propagate IR subtypes in lower_ir_node
This updates the type signatures for lower_ir_node to narrow the return
subtype where possible.

Previously, the return type of `lower_ir_node`'s first element was a generic
`IR`. However, the majority of `lower_ir_node` impelmentations return
the same subtype as the input. Of those that don't they typically return
just one or two subtypes (e.g. Scan returns a `Union` or `Scan`
depending on the input).

Only `Shuffle` was unable to be narrowed.

This ended up not being immediately useful to me, but I was already
mostly done updating things so I carried on.