[BUG] Incorrectly getting the caller's frame when searching for locals and globals in cudf.pandas (#18979)
## Description
<!-- Provide a standalone description of changes in this PR. -->
<!-- Reference any issues closed by this PR with "closes #1234". -->
<!-- Note: The pull request title will be included in the CHANGELOG. -->
This PR fixes a bug in how cudf.pandas retrieves `locals` and `globals`
when evaluating expressions with `DataFrame.query() `and
`DataFrame.eval()`. Previously, the implementation relied on a
hard-coded frame offset (`sys._getframe(level + 3)`), which failed with
`pytest-xdist` due to unpredictable call stack structure. We could not
rely on hard coding the offset `3`. This led to `NameErrors` when using
variables in expressions, because the variable could not be found in the
frame's local scope.

We now use `_find_user_frame()`, which searches the frame stack to find
the correct caller frame.

This PR also adds support for the global_dict parameter in
DataFrame.query, which was missing from the proxy method:

https://github.com/rapidsai/cudf/blob/594d26768ce86b9c2f389e851ae1afb77032c879/python/cudf/cudf/pandas/_wrappers/pandas.py#L1141-L1150
- [x] I am familiar with the [Contributing
Guidelines](https://github.com/rapidsai/cudf/blob/HEAD/CONTRIBUTING.md).
- [x] New or existing tests cover these changes.
- [x] The documentation is up to date with these changes.