Back to a plain function
This reverts the changes that returned a wrapper around IR.do_evaluate.

While looking at the nsys profiles I noticed we were now spending a lot
of time (~500 ms) in cloudpickle doing gerialization and deserialization.

The variant that used a wrapper wasn't serializable with plain pickle,
and the extra tricks that cloudpickle was doing apparently slowed things
down.

With this setup, we're back to 50ms for serialization and deserialization
of query 21 at SF 30.