Register Dask tokenizers for cudf_polars types.
This registers some custom normalizers for the most common types in the
task graph, to speed up Dask's tokenization. Previously, Dask had to fall
all the way back to the `pickle` tokenizer to tokenize these objects,
which is relatively slow.

I tested this locally on query 21 at SF 1K. On `main` tokenization took
about 1.3s. On this branch, it takes about 0.69s.