Avoid undefined behaviour in rolling_store_output_functor (#19069)
For rolling mean, the aggregation is computed by summing valid values in
the window and then dividing by the count in the store operation.

For duration types, the output type is also a duration (i.e. has an
integer representation), and so the division of the sum of the window by
the count is an integer division. Integer division by zero is undefined
behaviour so we must avoid it. If the compiler just emits the division
as is we might get a runtime trap, however, a sufficiently smart
compiler might deduce that, since count is initialized to zero and then
only incremented in the loop over the window bounds, and division by
zero is undefined, that the loop between start_index and end_index must
execute at least once (with a valid input value).

To avoid this, only store the output (also saving some memory traffic)
if the output is determined to be valid.

This might leave some uninitialized values in the output but by contract
you're not allowed to look at them because they are masked by a null.

---------

Co-authored-by: Bradley Dice <bdice@bradleydice.com>