Fix cudf_polars spilling
This fixes a bug where cudf_polars failed when using spilling.

https://github.com/rapidsai/cudf/pull/18970 made some assumptions about
the callables that we put in our task graph: namely that the first
argument is always some `IR.do_evaluate`. This is true, *except* for
when we post-process the task graph to do our spillable wrapping.

I'm hopeful that there's a better way to do this by integrating
spilling more into the IR, but this should be an OK fix in the meantime.

Closes https://github.com/rapidsai/cudf/issues/19100