Used TypeDict for CachingVisitor.state
This changes `CachingVisitor.state` to used a `TypedDict`. Previously,
we used a `Mapping[str, Any]`, which had two problems:

1. Risks typos in the key names causing unexpected KeyErrors
2. The `Any` type for the values of `state` mean that, at least by
   default, you won't get any type checking on things using values
   looked up from `state`.

Unit tests should catch all these, but this can provide some earlier
feedback on any issues.

The `total=False` keyword is used in the typeddict to indicate that
these keys are all optional. `mypy` *doesn't* error if you do a lookup
on an optional field of a TypedDict, so we do still need to handle
missing keys like we have been.