Optimize parquet reader's stats based row group filtering (#19156)
This PR adds optimizes parquet reader's row groups using stats by only building (min/max) stats for only the input columns that participate in the filter expression.

Before this PR, we used to build stats columns for all supported input columns regardless of whether they appeared in the filter expression or not. With this PR, we only add placeholder stats columns for unsupported and non-participating input columns. The change in this PR will also be applied to page pruning with stats (implemented in #18873) in another PR.

The biggest performance advantage from changes in this PR will be reflected in case of wide tables as well as in page pruning where we expand page level stats to row level stats.

Authors:
  - Muhammad Haseeb (https://github.com/mhaseeb123)

Approvers:
  - Shruti Shivakumar (https://github.com/shrshi)

URL: https://github.com/rapidsai/cudf/pull/19156