Add test for deserializing class instances
In https://github.com/rapidsai/cudf/pull/18953/files#r2167631601, we
found an issue with (de)serialization of cudf_polars data types with
a multi-GPU setup. Polars lets you use either datatype classes or
instances in most places. These are similar in most ways (hash the
same, compare equal) but our deserialization implementation relies on
isinstance checks which are not the same.

This could cause issues if we tried to deserialize a type like `'i8'`
that wasn't already in our cache. This simulates that situation on a
single worker by clearing the cache before deserialzation.

This was fixed in https://github.com/rapidsai/cudf/pull/19193, but
I've added a regression test here.