Basic `.profile` support in the streaming executor
This implements very basic support for `Timer` / `profile` in the
streaming executor for cudf-polars. It does the minimum amount necessary
to avoid the `TypeError` previously raised when calling `.profile()`.

In the documentation, we direct users toward alternatives (nsys for the
synchronous scheduler, dask's diagnostics for the distributed
scheduler).

xref https://github.com/rapidsai/cudf/issues/19253.