Move shuffle method defaulting to config options creation
Instead of setting the default shuffle method for the streaming
executor late, when lowering the task graph, instead set it when
creating our internal options object.

This will simplify tracking all configuration options for benchmark
runs, since we don't rely on this implicit default that is dependent
on the installation state at the time of the run.