Optimize object listing in pandas-tests diff CI (#19328)
This CI job lists all the objects under the `nightly` prefix and filters to the ones matching the filename query. The only keys that should match the query should be under `nightly/cudf/`, so update the prefix of the query to use that, which should speed things up a bit.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)
  - GALI PREM SAGAR (https://github.com/galipremsagar)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/19328