Improve error message for assert_column_eq
This changes our test-only assert_column_eq utility to provide a better
error message when running under pytest.

After we've detected that two columns are not equal, we'll cast the
columns to either python lists (for nested dtypes like lists or structs)
or numpy arrays.

We then use NumPy's or python's usual assertion machiner to compare.