Add `bounds_policy` to `pylibcudf.lists.segmented_gather` (#19411)
This adds a new `bounds_policy` keyword to `pylibucdf.lists.segmented_gather` that controls the behavior when performing out-of-bounds indexing on the individual list elements. The default is compatible with the previous behavior: `OutOfBoundsPolicy.DONT_CHECK`, where the value of the outputs is undefined. Specify `OutOfBoundsPolicy.NULLIFY` to set the value for out-of-bounds indexers to null.

Closes https://github.com/rapidsai/cudf/issues/19346

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Matthew Murray (https://github.com/Matt711)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/19411