Add a mode to co-process decompression and compression on host and device (#19203)
Added a HYBRID mode to decompression and compression. This mode potentially uses both CPU and GPU for these operations.

When enabled, inputs are sorted from the largest to the smallest; some number of leading blocks will be processed on the host. 
The selection of the point from which blocks are processed on the device depends on the total size of blocks assigned to host, and the size of the largest block assigned to device. Namely, the algorithm aims for the host to process `LIBCUDF_HOST_(DE)COMPRESSION_RATIO` times more data than the largest block that will be processed on the GPU. The default ratio is 100, and users can tune these values to control how much work will be done on the host. Larger value with lead to more work for the host.

This PR impacts the default engine because the inputs are now always sorted. Measured no impact on the benchmarks when the device (de)comp engine is used. We could potentially see (de)compression latency reduction for large files, but I haven't verified this.

TODO: expand tests to cover this mode

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Tianyu Liu (https://github.com/kingcrimsontianyu)
  - Muhammad Haseeb (https://github.com/mhaseeb123)

URL: https://github.com/rapidsai/cudf/pull/19203