Fix logic for number of unique values generated by data profile in benchmarks (#19540)
This PR clarifies and corrects the usage of cardinality in the data generator, and restores the previous distribution of the data generated for cardinality `c = 0` . When `c` is a non-negative integer less than the total number of values to be generated, then the number of unique elements in the output is at most `c`. Otherwise, if the cardinality is greater than or equal to the total number of values required, then the cardinality of the output is enforced i.e. the number of unique elements in the output is exactly `c`.

Authors:
  - Shruti Shivakumar (https://github.com/shrshi)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Muhammad Haseeb (https://github.com/mhaseeb123)

URL: https://github.com/rapidsai/cudf/pull/19540