Add hash-based SUM_WITH_OVERFLOW aggregation for INT64 values (#19403)
Contributes to #19243

This PR introduces a new aggregation kind, `SUM_WITH_OVERFLOW`, which returns a `STRUCT` containing the sum and a boolean indicating overflow. If an overflow occurs, the corresponding row in the overflow boolean column will be set to `true`.

Authors:
  - Yunsong Wang (https://github.com/PointKernel)

Approvers:
  - Robert (Bobby) Evans (https://github.com/revans2)
  - David Wendt (https://github.com/davidwendt)
  - Matthew Roeschke (https://github.com/mroeschke)
  - Nghia Truong (https://github.com/ttnghia)
  - Lawrence Mitchell (https://github.com/wence-)
  - Shruti Shivakumar (https://github.com/shrshi)

URL: https://github.com/rapidsai/cudf/pull/19403