Add object-oriented APIs for left semi- and anti- join (Part I) (#19778)
This PR introduces an object-oriented design for left semi- and anti- joins. Though the public `filtered_join` class enforces the reuse of the right table i.e. the filter table in this PR, the [follow-on PR](https://github.com/rapidsai/cudf/pull/19779) allows for reuse of either the left table or the right table in join operations. When the right table i.e. the filter table is reused, then the more memory-efficient and performant static_set data structure is used, wherein the right table is hashed and queried into for (non-)existence of each row in the left table for (anti-)semi-join. 
Note that reusing the right table matches the existing behavior of filtered joins, and is retained in the current design.

Authors:
  - Shruti Shivakumar (https://github.com/shrshi)

Approvers:
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)
  - Yunsong Wang (https://github.com/PointKernel)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/19778