Add MemoryResourceConfig to cudf-polars config
This adds a new configuration option to cudf-polars' config,
allowing users to specify the memory resource to create by default.

Currently, users either get the default behavior (typically a managed
memory resource) or can pass in a concrete memory resource.

In some cases, you might want to pass in a description of the memory
resource to use:

1. In our unit tests, we might want to specify a CudaAsyncMemoryResource
   with a relatively small initial pool size
2. In a distributed environment, you can't pass around concrete
   memory resource objects (which can't be serialized)