Avoid shadowing module names (#20071)
Using `plc` and `polars` as the member names of `DataType` inadvertently shadows the modules polars and plc for typing purposes. The Python interpreter is smart enough to not be tricked by this at runtime, but mypy is not. Since we import polars as pl, that one doesn't matter, but the `plc` attribute does matter and causes problems here. This global renaming avoids that issue.

Contributes to #17470

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Matthew Roeschke (https://github.com/mroeschke)

URL: https://github.com/rapidsai/cudf/pull/20071