Adjust rmm pool handling in PDSH benchmarks
This updates the handling of the RMM Pool argument for the PDSH CLI.
Previously, `--rmm-pool-size` was used for both sync and async RMM
memory resources.

As described in https://github.com/rapidsai/rmm/issues/2060, the meaning
of the pool size is different for these two types of memory resources,
and https://github.com/rapidsai/dask-cuda/pull/1563 is deprecating
passing a pool size to `dask_cuda.LocalCUDACluster` with
`rmm_async=True`.