Trace node execution in cudf-polars (#19895)
This PR introduces a new *low-overhead* tracing tool for cudf-polars. When enabled, we'll capture a record for each `IR.do_evaluate` node executed while running the polars query. The record will contain

1. The type of the IR node being executed
2. The number and size (in bytes and rows) of each input and output DataFrame
3. The RMM statistics before and after executing the node
4. The duration of the node execution.

The overhead of this when not tracing should just be a conditional check against a constant bool, and an extra function call. The overhead when tracing is enabled should be minimal (mostly getting the dataframe sizes, RMM statistics, and emitting the log record).

The implementation is based on structlog. This is helpful for us because we can inject additional context (like the PDSH query ID being run in our benchmarks) into the log records.

As an example of what this can be used for, I've updated our PDSH benchmarks to include a new `--collect-traces` flag. When set, this will include the `traces` on each `record` exported to our `pdsh_results.jsonl` file. The `records` for that run (one per query / iteration pair) will now contain a `list[traces]` field, with the traces from each node executed during each query.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Bradley Dice (https://github.com/bdice)
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/19895