Fix a race condition in the decode of delta encoded Parquet columns (#20216)
Issue #20187

Delta binary encoded columns ([reference](https://parquet.apache.org/docs/file-format/data-pages/encodings/#delta-encoding-delta_binary_packed--5)) are decoded one mini-block at a time, but the decoded values are written to the output in parallel with the decode of the next mini-block. Because of this, shared circular buffers are sized to `2 * max_delta_mini_block_size`. However, the first mini-block in each block encodes an additional element - the first value, stored in the header. Thus, the first mini-block can effectively contain `max_delta_mini_block_size + 1` elements. This additional element is not stored in the circular buffers, but the elements stored in the circular buffer are still stored in locations offset by one, to make the indexing simpler. Because of this offset, the last element of the first mini-block can (of a block, not of the file) be stored in the same location in the circular buffer as the first element of the previous mini-block (from the previous block). This is a race condition and can lead to incorrect values at these locations.

Example:
3 Blocks each having 2 mini blocks of 64 values:
B0b0 -> 1-64
B0b1 -> 65-127, 0
B1b0 -> 2-65 (since it's header value may have it skip 1) <= race at 65 (with B0b1)
B1b1 -> 66-127,0-1
B2b0 -> 3-66 (due to skip) <= race at 66 (with B1b1)
B2b1 -> 67-127,0-2

This PR expands the circular buffers to avoid the race condition.

PR fixes a race condition reported by `racecheck` when reading the repro file of #20187.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/20216