Cleanup parquet for simple columns (#19869)
This dramatically simplifies the parquet decode for simple columns (no nulls, no lists).  In this case we can directly copy the input to the output and don't need a buffer to map source -> destination. This is mainly a cleanup PR, it does not have an immediate performance impact.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/cudf/pull/19869