Fix parquet row number check for page bounds (#20217)
This fixes a bug in the row number checking that determines whether or not a page needs to be processed.  Before, if there are 20k rows per page and we skip 20k rows, page_start_row == max_row and we would still process the page despite there being nothing to do.  Ditto with the minimum row check.  Although this is a bug, it is effectively just a minor missed optimization opportunity so it didn't cause any tests to fail previously.  I've renamed the "max" variables to "end" to make the bug and the code clearer.

Authors:
  - Paul Mattione (https://github.com/pmattione-nvidia)

Approvers:
  - Muhammad Haseeb (https://github.com/mhaseeb123)
  - Shruti Shivakumar (https://github.com/shrshi)

URL: https://github.com/rapidsai/cudf/pull/20217