Use int64_t for the num_rows slot in parquet_reader_options
Although reading a single table of more than size_type::max() rows is not
possible, we may wish to specify options to read greater than
size_type::max() rows from a large parquet file and require that the
consumer of the options splits them up appropriately.

This is useful in chunked/streaming pipelines because we do not have to
recreate the read_parquet options API and can instead get the caller to
provide an options object that we pick apart.

Internally the reader implementation already uses int64 for num_rows, so
this is really just a (breaking) interface change.