Add context to IR.do_evaluate
This adds a keyword-only `context` argument to cudf_polars
IR.do_evaluate method. The purpose to provide access to special pieces
of data that might be necessary for controlling an IR nodes' execution,
but doesn't belong on the IR node itself as a non-child argument.
Specifically, we'd like to provide a CUDA `stream` argument, but we
generalize that slightly and provide a system for providing arbitrary
data.