Add context to IR.do_evaluate (#20322)
This adds a keyword-only `context` argument to cudf_polars IR.do_evaluate method. The purpose to provide access to special pieces of data that might be necessary for controlling an IR nodes' execution, but doesn't belong on the IR node itself as a non-child argument. Specifically, we'd like to provide a CUDA `stream` argument as part of https://github.com/rapidsai/cudf/issues/20228, but we generalize that slightly and provide a system for providing arbitrary data.

A few notes on the implementation:

- For now, the context is just an empty dataclass. I suspect its design might change in the future.
- I've opted to push the creation of the context as high as possible. For now it's created in `_callback` and passed into `ir.evaluate` / `evaluate_streaming` and from there to all the methods that require it.
- There's some awkwardness between how our IR nodes and Dask's task graph treat arguments. I've opted to make `context` keyword only in `IR.do_evaluate(..., context)`. However, Dask's task graph doesn't really deal with that. It wants a tuple of `(function, arg1, arg2, ...)`. So that requires using `functools.partial(function, context=context)(arg1, arg2, ...)`.
- After implementing this, I realized that `Expr.evaluate` *also* takes a context, and its a different type `ExecutionContext` :( I can rename the IR variant if we want.

Just a draft for now, and probably not worth reviewing until I have a branch somewhere that combines CUDA streams with this to verify it meets our needs.

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/20322