Defer offsets_to_windows calls
This defers the point at which we call offsets_to_windows, until a time
when we have a CUDA stream in do_evaluate.

We do still perform some of the work (calling duration_to_int) early on
to catch any translation issues (e.g. we want to raise with month is
non-zero, since we don't support that).