Add CUDAStreamPolicy to cudf-polars configuration (#20366)
This adds a new configuration option to cudf-polars that enables control over how new CUDA streams are created (notably when evaluating Scan and DataFrameScan nodes).

The default behavior is compatible with the previous behavior: everything uses the defualt stream. We provide 2 policies:

- default: use the default CUDA stream
- new: always create a new CUDA stream

In the future, we'll add a "pool" policy. We'll make it available when using the rapidsmpf runtime, which does have a pool available to use.

The implementation uses this policy to create the 0-argument `get_cuda_stream` callable that's attached to IRExecutionContext. We've updated the handful of spots that previously used
`cudf_polars.cuda_utils.get_cuda_stream()` to use
`context.get_cuda_stream()` (which, by default, will be unchanged).

Closes https://github.com/rapidsai/cudf/issues/20337
Closes https://github.com/rapidsai/cudf/issues/20365

Authors:
  - Tom Augspurger (https://github.com/TomAugspurger)

Approvers:
  - Richard (Rick) Zamora (https://github.com/rjzamora)
  - Matthew Murray (https://github.com/Matt711)

URL: https://github.com/rapidsai/cudf/pull/20366