[Java] Add optional native deps loader (#20414)
Relates to https://github.com/NVIDIA/spark-rapids-jni/issues/3891

Adds a `loadOptionalNativeDeps` function to `NativeDepsLoader.java` which can be used to try to load native deps which might not always be available on the system. The motivating use case is https://github.com/NVIDIA/spark-rapids-jni/pull/3899 which uses it to load a libnvmljni.so which depends on libnvidia-ml.so on the host which might not be present, so the linking could fail, in which case we want to be able to do a graceful fallback.

Authors:
  - Zach Puller (https://github.com/zpuller)

Approvers:
  - Alessandro Bellina (https://github.com/abellina)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/20414