Add option to reset memory resource in PDSH
This adds a CLI flag to reset the memory pool between iterations in the
PDSH benchmark. The default is to reset the MR, which is a change in
behavior from previous behavior.

We really only expect this to have an effect for the default memory
resource, which uses UVM. We've observed some instability in memory
usage, which this flag might help address.