Handle dealloc in stream-ordered cudf-polars ops
This updates cudf-polars' usage of CUDA streams to safely handle
deallocation.

Consider the following sequnce of operations:

1. Read some data on stream A
2. Read some data on stream B
3. Concat data from A and B on new stream C

We currently ensure that C is downstream of `A` and `B` before doing the
concat. But then our execution will typically drop all references to the
data on A and B, at which point Python's reference counting will cause
the data there to be deallocated, which will call *stream ordered*
deallocation on the original stream.

We need to ensure that this stream ordered deallocation happens after
the result from 3 (on stream C) is done, and so we join stream C into
each of A and B.