Expose java GatherMap internals and add toString to AST (#20483)
This does two small things. If you want them in separate PRs I can do that, but they are small.

This allows a GatherMap to be processed through JNI in a non-destructive way. It lets us get access to the underlying buffer without turning it into a `ColumnView` or breaking it apart.

This also adds in some debugging for AST expressions to help understand what is going on before they are compiled.

I have added no tests because the code changes are so simple.

Authors:
  - Robert (Bobby) Evans (https://github.com/revans2)

Approvers:
  - Alessandro Bellina (https://github.com/abellina)

URL: https://github.com/rapidsai/cudf/pull/20483