Use a lower bound when estimating the partial file-size (#20193)
While working on https://github.com/rapidsai/cudf/pull/20161, I discovered that cudf-polars can significantly underestimate the in-memory size of a Parquet string column. In order to estimate the in-memory size, we currently use column-chunk metadata from the Parquet file. This metadata stores the uncompressed byte size of each column. However, the uncompressed size may still include **dictionary encoding**. This means the metadata value can be significantly smaller than the real in-memory footprint. This is especially true when the column has a small unique-value count.

There are many possible ways to improve our partial file-size estimate. This PR takes a relatively simple approach: We set a lower-bound on the partial file-size estimate. This lower bound simply corresponds to the size of a single element (e.g. 8 bytes for strings) and multiplies that value by the estimated average row-count in a file.

**NOTE**: This change will probably degrade the performance of some PDS-H queries, because it will result in a larger partition count for dictionary-encoded data. For example, our partition size for Query 1 is currently very aggressive - It is roughly ~5x larger than we are expecting from the `blocksize` argument. The flip side of this issue is that we can probably handle a slightly-larger blocksize configuration (or a higher spilling limit) with the changes in this PR.

Authors:
  - Richard (Rick) Zamora (https://github.com/rjzamora)

Approvers:
  - Tom Augspurger (https://github.com/TomAugspurger)

URL: https://github.com/rapidsai/cudf/pull/20193